Split RVC4 specific test case into separate script node test
